feature: Refactoring to account for Factions/non-mob AIs better.
The concept of a Pawn is now generic; a mob AI drives a mob/atom Pawn, a faction AI drives a faction datum Pawn, a director AI would drive an Event Pawn or sth.
Reorganized Actions - split into subdirs where the Action only makes sense for a subtype.
Fixed a ton of nullchecks to be typechecks instead.