feature: (1) Saner trade prices. feature: (2) Binary Heap-based priority queue
(1) Last unit prices are stored in Yet Another Global Array and seeded from a JSON.
The trade AI incorporates this into their guesstimates where the last price is better than their guess, adjusting the pricing.
This stabilizes the price offers somewhat, as long as the AI needs are not extreme - a desperate AI will still make desperate deals.
Got a basic two-faction setup going! Still needs production/consumption implemented.

(2) This fixes a LOOOOOOONG-standing bad implementation of PQs in SS13.
The original implementation is O(N) worst-case on insertion due to swapping all tail items to the next slot.
The PQ is defined as an abstract class with a legacy OrderedList impl and a new BinHeap impl, with a define to switch between them.
The legacy implementation uses a strange convention that had been retained for now in both classes for integration's sake.